requirements.txt
Flask==3.0.1
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
flask-restx==1.3.0
mysqlclient==2.2.1
flask-cors==4.0.0
python-dotenv==1.0.1
python-dateutil==2.8.2
gunicorn==21.2.0
redis==5.0.1
